test(e2e): move websocket E2E into console.test and remove standalone ws.test to avoid server start races
